third_party/blink/web_tests/dom/mutation-event-tests/fast/events/mutation-during-replace-child-2-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Ensures that replaceChild() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS target.replaceChild(newChild, oldChild); threw exception HierarchyRequestError: Failed to execute 'replaceChild' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
